Rename clinician View Info etc. to be descriptive
