+++ src/navigation/navigation.js
@@ -4,8 +4,8 @@ import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import LoginScreen from '../screens/LoginScreen';
 import ClinicianHomeScreen from '../screens/ClinicianHomeScreen';
 import PatientHomeScreen from '../screens/PatientHomeScreen';
-import ClinicianViewScreen from '../screens/ClinicianViewScreen';
-import ClinicianViewInfoScreen from '../screens/ClinicianViewInfoScreen';
+import PatientsScreen from '../screens/PatientsScreen';
+import PatientScreen from '../screens/PatientScreen';
 import PatientMessagesScreen from '../screens/PatientMessagesScreen';
 import PatientUploadScreen from '../screens/PatientUploadScreen';
 
@@ -17,8 +17,8 @@ export default function MainNavigator() {
         <Stack.Navigator>
             <Stack.Screen name="Login" component={LoginScreen}  />
             <Stack.Screen name="Clinician Home" component={ClinicianHomeScreen} />
-            <Stack.Screen name="Patients" component={ClinicianViewScreen} />
-            <Stack.Screen name="Patient" component={ClinicianViewInfoScreen} />
+            <Stack.Screen name="Patients" component={PatientsScreen} />
+            <Stack.Screen name="Patient" component={PatientScreen} />
             <Stack.Screen name="PatientHome" component={PatientHomeScreen} />
             <Stack.Screen name="PatientUpload" component={PatientUploadScreen} />
             <Stack.Screen name="PatientMessages" component={PatientMessagesScreen} />
