+++ src/screens/PatientScreen.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { StyleSheet, Text, View, Button } from 'react-native';
-import ClinicianViewScreen from './ClinicianViewScreen';
+import ClinicianViewScreen from './PatientsScreen';
 import MainNavigator from'../navigation/navigation';
 
 const ClinicianViewInfoScreen = ({navigation}) => {
